+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# v1.0.1
+## 09/27/2022
+
+1. [](#new)
+    * added `onMailerSendVars`, `onMailerSendBeforeSend` and `onMailerSendAfterSend` events
+    * added support for **arrays** and also **comma-separated strings** for multiple addresses
+    * added support for `Your Name <hello@yoursite.com>`, `{hello@yoursite.com: Your Name}` and `'<hello@yoursite.com>'` email formats
+
 # v1.0.0
 ## 09/26/2022